[takea-look#31] package명 변경 : my.takealook
