+++ app/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 
 android {
     defaultConfig {
-        applicationId = "com.easternkite.takealook"
+        applicationId = "my.takealook"
         targetSdk = 35
         versionCode = 1
         versionName = "1.0"
