--- app/src/main/java/com/easternkite/takealook/TklApplication.kt
@@ -1,7 +1,0 @@
-package com.easternkite.takealook
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TklApplication : Application()
