+++ app/src/main/java/my/takealook/TklApplication.kt
@@ -1,0 +1,7 @@
+package my.takealook
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TklApplication : Application()
