+++ build-logic/convention/src/main/kotlin/AndroidApplicationComposeConventionPlugin.kt
@@ -1,8 +1,8 @@
 import com.android.build.api.dsl.ApplicationExtension
-import com.easternkite.takealook.configureAndroidCompose
-import com.easternkite.takealook.configureHilt
-import com.easternkite.takealook.configureKotlinAndroid
-import com.easternkite.takealook.configureNavigation3
+import my.takealook.configureAndroidCompose
+import my.takealook.configureHilt
+import my.takealook.configureKotlinAndroid
+import my.takealook.configureNavigation3
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
