--- build-logic/convention/src/main/kotlin/com/easternkite/takealook/AndroidCompose.kt
@@ -1,25 +1,0 @@
-package com.easternkite.takealook
-
-import com.android.build.api.dsl.CommonExtension
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.dependencies
-
-internal fun Project.configureAndroidCompose(
-    commonExtension: CommonExtension<*, *, *, *, *, *>,
-) {
-    commonExtension.apply {
-        buildFeatures {
-            compose = true
-        }
-
-        dependencies {
-            val bom = libs.findLibrary("androidx-compose-bom").get()
-            add("implementation", platform(bom))
-            add("implementation", libs.findLibrary("androidx-ui").get())
-            add("implementation", libs.findLibrary("androidx-ui-tooling-preview").get())
-            add("implementation", libs.findLibrary("androidx-material3").get())
-            add("implementation", libs.findLibrary("androidx-activity-compose").get())
-            add("debugImplementation", libs.findLibrary("androidx-ui-tooling").get())
-        }
-    }
-}
