--- build-logic/convention/src/main/kotlin/com/easternkite/takealook/Libs.kt
@@ -1,9 +1,0 @@
-package com.easternkite.takealook
-
-import org.gradle.api.Project
-import org.gradle.api.artifacts.VersionCatalog
-import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.kotlin.dsl.getByType
-
-val Project.libs
-    get(): VersionCatalog = extensions.getByType<VersionCatalogsExtension>().named("libs")
