+++ build-logic/convention/src/main/kotlin/my/takealook/CoilAndroid.kt
@@ -1,0 +1,11 @@
+package my.takealook
+
+import org.gradle.api.Project
+
+fun Project.configureCoil() {
+    val coilCompose = libs.findLibrary("coil-compose").get()
+    dependencies.add("implementation", coilCompose)
+
+    val coilNetworkOkhttp = libs.findLibrary("coil-network-okhttp").get()
+    dependencies.add("implementation", coilNetworkOkhttp)
+}
