+++ build-logic/convention/src/main/kotlin/my/takealook/HiltAndroid.kt
@@ -1,0 +1,14 @@
+package my.takealook
+
+import org.gradle.api.Project
+
+fun Project.configureHilt() {
+    with(pluginManager) {
+        apply("com.google.dagger.hilt.android")
+        apply("com.google.devtools.ksp")
+    }
+    val hilt = libs.findLibrary("hilt-android").get()
+    dependencies.add("implementation", hilt)
+    val hiltCompiler = libs.findLibrary("hilt-compiler").get()
+    dependencies.add("ksp", hiltCompiler)
+}
