+++ core/designsystem/src/main/kotlin/my/takealook/theme/Color.kt
@@ -1,0 +1,98 @@
+package my.takealook.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Primary0 = Color(0xFF000000)
+val Primary10 = Color(0xFF002018)
+val Primary20 = Color(0xFF00382B)
+val Primary30 = Color(0xFF005140)
+val Primary40 = Color(0xFF1E6A57)
+val Primary50 = Color(0xFF3C836F)
+val Primary60 = Color(0xFF579E88)
+val Primary70 = Color(0xFF72B9A2)
+val Primary80 = Color(0xFF8DD5BD)
+val Primary90 = Color(0xFFA8F1D8)
+val Primary95 = Color(0xFFB9FFE7)
+val Primary99 = Color(0xFFF3FFF9)
+val Primary100 = Color(0xFFFFFFFF)
+
+val Secondary0 = Color(0xFF000000)
+val Secondary10 = Color(0xFF101E1A)
+val Secondary20 = Color(0xFF25332E)
+val Secondary30 = Color(0xFF3B4A44)
+val Secondary40 = Color(0xFF52625C)
+val Secondary50 = Color(0xFF6B7A74)
+val Secondary60 = Color(0xFF84948D)
+val Secondary70 = Color(0xFF9FAFA8)
+val Secondary80 = Color(0xFFBACAC3)
+val Secondary90 = Color(0xFFD6E6DE)
+val Secondary95 = Color(0xFFE4F5ED)
+val Secondary99 = Color(0xFFF3FFF9)
+val Secondary100 = Color(0xFFFFFFFF)
+
+val Tertiary0 = Color(0xFF000000)
+val Tertiary10 = Color(0xFF081E29)
+val Tertiary20 = Color(0xFF1F333F)
+val Tertiary30 = Color(0xFF354956)
+val Tertiary40 = Color(0xFF4D616E)
+val Tertiary50 = Color(0xFF657A87)
+val Tertiary60 = Color(0xFF7F94A2)
+val Tertiary70 = Color(0xFF99AEBD)
+val Tertiary80 = Color(0xFFB4C9D8)
+val Tertiary90 = Color(0xFFD0E6F5)
+val Tertiary95 = Color(0xFFE3F3FF)
+val Tertiary99 = Color(0xFFFBFCFF)
+val Tertiary100 = Color(0xFFFFFFFF)
+
+val Error0 = Color(0xFF000000)
+val Error10 = Color(0xFF410002)
+val Error20 = Color(0xFF690005)
+val Error30 = Color(0xFF93000A)
+val Error40 = Color(0xFFBA1A1A)
+val Error50 = Color(0xFFDE3730)
+val Error60 = Color(0xFFFF5449)
+val Error70 = Color(0xFFFF897D)
+val Error80 = Color(0xFFFFB4AB)
+val Error90 = Color(0xFFFFDAD6)
+val Error95 = Color(0xFFFFEDEA)
+val Error99 = Color(0xFFFFFBFF)
+val Error100 = Color(0xFFFFFFFF)
+
+val Neutral0 = Color(0xFF000000)
+val Neutral4 = Color(0xFF1A1C1B)
+val Neutral6 = Color(0xFF0F1512)
+val Neutral10 = Color(0xFF1A1C1B)
+val Neutral12 = Color(0xFF1B211E)
+val Neutral17 = Color(0xFF252B29)
+val Neutral20 = Color(0xFF2F3130)
+val Neutral22 = Color(0xFF303633)
+val Neutral24 = Color(0xFF343B38)
+val Neutral30 = Color(0xFF454746)
+val Neutral40 = Color(0xFF5D5F5D)
+val Neutral50 = Color(0xFF767776)
+val Neutral60 = Color(0xFF90918F)
+val Neutral70 = Color(0xFFAAABA9)
+val Neutral80 = Color(0xFFC6C7C4)
+val Neutral87 = Color(0xFFD5DBD7)
+val Neutral90 = Color(0xFFE2E3E0)
+val Neutral92 = Color(0xFFE3EAE5)
+val Neutral94 = Color(0xFFE9EFEB)
+val Neutral95 = Color(0xFFF1F1EE)
+val Neutral96 = Color(0xFFEFF5F1)
+val Neutral98 = Color(0xFFF5FBF6)
+val Neutral99 = Color(0xFFFCFCFA)
+val Neutral100 = Color(0xFFFFFFFF)
+
+val NeutralVariant0 = Color(0xFF000000)
+val NeutralVariant10 = Color(0xFF171D1B)
+val NeutralVariant20 = Color(0xFF2C322F)
+val NeutralVariant30 = Color(0xFF424845)
+val NeutralVariant40 = Color(0xFF5A605D)
+val NeutralVariant50 = Color(0xFF737875)
+val NeutralVariant60 = Color(0xFF8C928F)
+val NeutralVariant70 = Color(0xFFA7ADA9)
+val NeutralVariant80 = Color(0xFFC2C8C4)
+val NeutralVariant90 = Color(0xFFDEE4E0)
+val NeutralVariant95 = Color(0xFFEDF2EE)
+val NeutralVariant99 = Color(0xFFF8FEF9)
+val NeutralVariant100 = Color(0xFFFFFFFF)
