+++ core/designsystem/src/main/kotlin/my/takealook/theme/Theme.kt
@@ -1,0 +1,110 @@
+package my.takealook.theme
+
+import android.os.Build
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+private val DarkColorScheme = darkColorScheme(
+    primary = Primary80,
+    onPrimary = Primary20,
+    primaryContainer = Primary30,
+    onPrimaryContainer = Primary90,
+    secondary = Secondary80,
+    onSecondary = Secondary20,
+    secondaryContainer = Secondary30,
+    onSecondaryContainer = Secondary90,
+    tertiary = Tertiary80,
+    onTertiary = Tertiary20,
+    tertiaryContainer = Tertiary30,
+    onTertiaryContainer = Tertiary90,
+    surfaceDim = Neutral6,
+    surface = Neutral6,
+    onSurface = Neutral90,
+    surfaceVariant = NeutralVariant30,
+    onSurfaceVariant = NeutralVariant80,
+    surfaceBright = Neutral24,
+    surfaceContainerLowest = Neutral4,
+    surfaceContainerLow = Neutral10,
+    surfaceContainer = Neutral12,
+    surfaceContainerHigh = Neutral17,
+    surfaceContainerHighest = Neutral22,
+    inversePrimary = Primary40,
+    inverseOnSurface = Neutral20,
+    inverseSurface = Neutral90,
+    error = Error80,
+    onError = Error20,
+    errorContainer = Error30,
+    onErrorContainer = Error90,
+    outline = NeutralVariant60,
+    outlineVariant = NeutralVariant30,
+    scrim = Neutral0,
+    background = Neutral10,
+)
+
+private val LightColorScheme = lightColorScheme(
+    primary = Primary40,
+    onPrimary = Primary100,
+    primaryContainer = Primary90,
+    onPrimaryContainer = Primary30,
+    secondary = Secondary40,
+    onSecondary = Secondary100,
+    secondaryContainer = Secondary90,
+    onSecondaryContainer = Secondary30,
+    tertiary = Tertiary40,
+    onTertiary = Tertiary100,
+    tertiaryContainer = Tertiary90,
+    onTertiaryContainer = Tertiary30,
+    surfaceDim = Neutral87,
+    surface = Neutral98,
+    onSurface = Neutral10,
+    surfaceVariant = NeutralVariant90,
+    onSurfaceVariant = NeutralVariant30,
+    surfaceBright = Neutral98,
+    surfaceContainerLowest = Neutral100,
+    surfaceContainerLow = Neutral96,
+    surfaceContainer = Neutral94,
+    surfaceContainerHigh = Neutral92,
+    surfaceContainerHighest = Neutral90,
+    inversePrimary = Primary80,
+    inverseOnSurface = Neutral95,
+    inverseSurface = Neutral20,
+    error = Error40,
+    onError = Error100,
+    errorContainer = Error90,
+    onErrorContainer = Error30,
+    outline = NeutralVariant50,
+    outlineVariant = NeutralVariant80,
+    scrim = Neutral0,
+    background = Neutral99,
+    onBackground = Neutral10,
+)
+
+@Composable
+fun TklTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    // Dynamic color is available on Android 12+
+    dynamicColor: Boolean = false,
+    content: @Composable () -> Unit
+) {
+    val colorScheme = when {
+        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+            val context = LocalContext.current
+            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+        }
+
+        darkTheme -> DarkColorScheme
+        else -> LightColorScheme
+    }
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        typography = Typography,
+        content = content
+    )
+}
