--- core/network/src/main/kotlin/com/easternkite/takealook/model/StickerCategoryResult.kt
@@ -1,7 +1,0 @@
-package com.easternkite.takealook.model
-
-data class StickerCategoryResult(
-    val id: Int? = null,
-    val name: String,
-    val thumbnailUrl: String
-)
