--- core/network/src/main/kotlin/com/easternkite/takealook/model/StickerResult.kt
@@ -1,9 +1,0 @@
-package com.easternkite.takealook.model
-
-data class StickerResult(
-    val id: Int? = null,
-    val name: String,
-    val iconUrl: String,
-    val thumbnailUrl: String,
-    val categoryId: Int
-)
