+++ core/network/src/main/kotlin/my/takealook/model/StickerCategoryResult.kt
@@ -1,0 +1,7 @@
+package my.takealook.model
+
+data class StickerCategoryResult(
+    val id: Int? = null,
+    val name: String,
+    val thumbnailUrl: String
+)
