+++ core/network/src/main/kotlin/my/takealook/model/StickerResult.kt
@@ -1,0 +1,9 @@
+package my.takealook.model
+
+data class StickerResult(
+    val id: Int? = null,
+    val name: String,
+    val iconUrl: String,
+    val thumbnailUrl: String,
+    val categoryId: Int
+)
