+++ gradle/libs.versions.toml
@@ -1,29 +1,32 @@
 [versions]
 # Android Project Settings
-namespace = "com.easternkite.takealook"
+namespace = "my.takealook"
 compileSdk = "36"
 minSdk = "24"
 
-agp = "8.9.3"
-androidTools = "31.10.1"
-kotlin = "2.1.21"
+agp = "8.11.1"
+androidTools = "31.11.1"
+kotlin = "2.2.0"
 coreKtx = "1.16.0"
 junit = "4.13.2"
 junitVersion = "1.2.1"
 espressoCore = "3.6.1"
-lifecycleRuntimeKtx = "2.9.1"
+lifecycleRuntimeKtx = "2.9.2"
 activityCompose = "1.10.1"
 composeBom = "2025.06.00"
 hilt = "2.56.2"
 ksp = "2.1.21-2.0.2"
 coil = "3.2.0"
 retrofit = "2.11.0"
+composeBom = "2025.07.00"
+hilt = "2.57"
+ksp = "2.2.0-2.0.2"
+coil = "3.3.0"
 okhttp = "4.12.0"
 
-nav3Core = "1.0.0-alpha03"
-lifecycleViewmodelNav3 = "1.0.0-alpha01"
-kotlinSerialization = "2.1.21"
-kotlinxSerializationCore = "1.8.1"
+nav3Core = "1.0.0-alpha05"
+lifecycleViewmodelNav3 = "1.0.0-alpha03"
+kotlinxSerializationCore = "1.9.0"
 material3AdaptiveNav3 = "1.0.0-SNAPSHOT"
 
 [libraries]
